% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_smoother_ar1.R
\name{precision_sampler_ar1}
\alias{precision_sampler_ar1}
\title{Samples random numbers from an \eqn{N}-variate normal distribution
specified by the precision matrix \eqn{P} and location vector \eqn{L} as in:
\deqn{N(P^{-1}L, P^{-1})}
where the precision matrix \eqn{P} is bi-diagonal with the diagonal elements 
given in the vector argument \code{precision_diag} and the off-diagonal element 
is given in the scalar argument \code{precision_offdiag}, and the location 
vector \eqn{L} is provided in the vector argument \code{location}.}
\usage{
precision_sampler_ar1(precision_diag, precision_offdiag, location)
}
\arguments{
\item{precision_diag}{an \eqn{N}-vector with the diagonal elements of the 
precision matrix \eqn{P}}

\item{precision_offdiag}{a numeric scalar with the off-diagonal element of 
the precision matrix \eqn{P}}

\item{location}{an \eqn{N}-vector with the location parameter \eqn{L}}
}
\value{
an \eqn{N}-vector with random draws from the multivariate normal 
distribution
}
\description{
Samples random numbers from an \eqn{N}-variate normal distribution
specified by the precision matrix \eqn{P} and location vector \eqn{L} as in:
\deqn{N(P^{-1}L, P^{-1})}
where the precision matrix \eqn{P} is bi-diagonal with the diagonal elements 
given in the vector argument \code{precision_diag} and the off-diagonal element 
is given in the scalar argument \code{precision_offdiag}, and the location 
vector \eqn{L} is provided in the vector argument \code{location}.
}
