% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig2.R
\name{rig2}
\alias{rig2}
\title{Sample from the inverted gamma-2 distribution}
\usage{
rig2(n, s, nu)
}
\arguments{
\item{n}{a positive integer giving the number of draws.
\strong{C++}: an \code{int} object.}

\item{s}{a positive real scalar giving the scale parameter.
\strong{C++}: a \code{double} object.}

\item{nu}{a positive real scalar giving the degrees-of-freedom parameter.
\strong{C++}: a \code{double} object.}
}
\value{
A numeric vector of length \code{n} containing independent draws
from the inverted gamma-2 distribution. \strong{C++}: an
\code{arma::vec} object.
}
\description{
Samples independent random numbers from the inverted gamma-2
distribution, denoted by \eqn{IG2(s, \nu)}, with density
\deqn{f(x) = \frac{(s/2)^{\nu/2}}{\Gamma(\nu/2)}
x^{-(\nu+2)/2}\exp\left(-\frac{s}{2x}\right), \quad x > 0.}
The draws are generated as \eqn{s/Z}, where \eqn{Z} follows a chi-squared
distribution with \eqn{\nu} degrees of freedom.

This distribution is commonly used for variance parameters in Bayesian
models.
}
\details{
This function is based on C++ code from the GPL-3 R package
\pkg{bsvars} by Woźniak (2024, 2025). It uses random-number generators from
the \pkg{armadillo} library by Sanderson & Curtin (2025), made available to R
through the \pkg{RcppArmadillo} package by Eddelbuettel et al. (2025).
}
\examples{
rig2(5, 1, 10)

}
\references{
Bauwens L., Lubrano M., Richard J.-F. (1999). \emph{Bayesian Inference in
Dynamic Econometric Models}. Oxford University Press.

Eddelbuettel D., François R., Bates D., Ni B., Sanderson C. (2025).
RcppArmadillo: 'Rcpp' Integration for the 'Armadillo' Templated Linear
Algebra Library. R package version 15.0.2-2.
\doi{10.32614/CRAN.package.RcppArmadillo}

Sanderson C., Curtin R. (2025). Armadillo: An Efficient Framework for
Numerical Linear Algebra. International Conference on Computer and
Automation Engineering, 303--307. \doi{10.1109/ICCAE64891.2025.10980539}

Woźniak T. (2024). bsvars: Bayesian Estimation of Structural Vector
Autoregressive Models, R package version 3.2.
\doi{10.32614/CRAN.package.bsvars}

Woźniak T. (2025). Fast and Efficient Bayesian Analysis of Structural
Vector Autoregressions Using the R Package bsvars. University of Melbourne
Working Paper, 1--25. \doi{10.48550/arXiv.2410.15090}
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
