% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_smoother_ar1.R
\name{precision_sampler_ar1}
\alias{precision_sampler_ar1}
\title{Samples draws from a multivariate normal distribution using the 
precision sampler by}
\usage{
precision_sampler_ar1(precision_diag, precision_offdiag, location)
}
\arguments{
\item{precision_diag}{an \eqn{N}-vector with the diagonal elements of the 
precision matrix \eqn{P}}

\item{precision_offdiag}{a numeric scalar with the off-diagonal element of 
the precision matrix \eqn{P}}

\item{location}{an \eqn{N}-vector with the location parameter \eqn{L}}
}
\value{
an \eqn{N}-vector with random draws from the multivariate normal 
distribution
}
\description{
Samples random numbers from an \eqn{N}-variate normal distribution
specified by the precision matrix \eqn{P} and location vector \eqn{L} as in:
\deqn{N(P^{-1}L, P^{-1})}
where the precision matrix \eqn{P} is bi-diagonal with the diagonal elements 
given in the vector argument \code{precision_diag} and the off-diagonal element 
is given in the scalar argument \code{precision_offdiag}, and the location 
vector \eqn{L} is provided in the vector argument \code{location}.
}
\details{
This function is based on C++ code from the R package \pkg{stochvol}
by Hosszejni D., Kastner G. (2025) and Kastner G. (2016).
}
\references{
Chan J.C.C., Jeliazkov I. (2009). Efficient simulation and integrated 
likelihood estimation in state space models. International Journal of 
Mathematical Modelling and Numerical Optimisation, 
1(1/2), <doi:10.1504/IJMMNO.2009.030090>.

Kastner G. (2016). Dealing with Stochastic Volatility in Time Series Using 
the R Package stochvol. Journal of Statistical Software, 69(5), 1–30. 
<doi:10.18637/jss.v069.i05>.

Hosszejni D., Kastner G. (2025). stochvol: Efficient Bayesian Inference for 
Stochastic Volatility (SV) Models. R package version 3.2.8, 
<doi:10.32614/CRAN.package.stochvol>
}
\author{
Tomasz Woźniak \email{wozniak.tom@pm.me}
}
