% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StealLikeBayes-package.R
\docType{package}
\name{StealLikeBayes-package}
\alias{StealLikeBayes-package}
\alias{StealLikeBayes}
\title{StealLikeBayes: What the Package Does (Short Line)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line).
}
\keyword{internal}
