% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StealLikeBayes-package.R
\docType{package}
\name{StealLikeBayes-package}
\alias{StealLikeBayes-package}
\alias{StealLikeBayes}
\title{A C++ compendium of Bayesian statistical routines for R packages}
\description{
This is a compendium of C++ routines useful for Bayesian 
statistics. We steal other people's C++ code, repurpose it, and export it so 
developers of R packages can use it in their C++ code. We actually don't 
steal anything, or claim that Thomas Bayes did, but copy code that is 
compatible with our GPL 3 licence, fully acknowledging the authorship of the 
original code.
}
\note{
This is a collaborative project. You are welcome to contribute. Please,
read the documentation at \url{https://bsvars.org/StealLikeBayes/dev/} and 
StealLikeBayes!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bsvars.org/StealLikeBayes/}
  \item Report bugs at \url{https://github.com/bsvars/StealLikeBayes/issues}
}

}
\author{
\strong{Maintainer}: Tomasz Woźniak \email{wozniak.tom@pm.me} (\href{https://orcid.org/0000-0003-2212-2378}{ORCID}) (contributions: rnorm1_precision_sampler, rgennorm)

Authors:
\itemize{
  \item Xiaolei Wang \email{adamwang15@gmail.com} (\href{https://orcid.org/0009-0005-6192-9061}{ORCID}) (contributions: rhaar1)
  \item Longcan Li \email{longcando@outlook.com} (\href{https://orcid.org/0009-0007-2190-768X}{ORCID}) (contributions: sample_variances_horseshoe)
  \item Filip Reierson \email{filip.reierson@gmail.com} (\href{https://orcid.org/0009-0003-2393-5810}{ORCID}) (contributions: rtmvnorm)
  \item Jianying Shelly Xie \email{shellyyinggxie@gmail.com} (\href{https://orcid.org/0009-0008-6451-6416}{ORCID}) (contributions: sample_variances_normal_gamma)
}

}
\keyword{Bayesian}
\keyword{internal}
