% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StealLikeBayes-package.R
\docType{package}
\name{StealLikeBayes-package}
\alias{StealLikeBayes-package}
\alias{StealLikeBayes}
\title{A C++ compendium of Bayesian statistical routines for R packages}
\description{
This is a compendium of C++ routines useful for Bayesian 
statistics. We steal other people's C++ code, repurpose it, and export it so 
developers of R packages can easily use it in their C++ code. We actually 
don't steal anything, but copy code that is compatible with our GPL 3 
licence, fully acknowledging the authorship of the original code.
}
\note{
This is a collaborative project. You are welcome to contribute. Please,
read the documentation and get to work of stealing like Bayes!
}
\author{
\strong{Maintainer}: Tomasz Woźniak \email{wozniak.tom@pm.me} (\href{https://orcid.org/0000-0003-2212-2378}{ORCID})

}
\keyword{internal}
